feat(#487): add guideline list to drug detail
